KV 2.0: split kv_planner.py into kv_stream/kv_cache/kv_planner, rename for clarity
kv_planner.py had grown to 1233 lines mixing four concerns: stream
declaration, the KVCache plan/pool object, the page-fitting algorithm, and
debug tooling. Split along those lines:

- kv_stream.py (~200 lines): KVMode/KVSpec/KVStream/FlatStream and the
  merge-identical-streams logic -- the declaration vocabulary a model
  builder actually writes.
- kv_cache.py (~500 lines): the KVCache class plus its direct dependents
  (KVGroupConfig, pages_per_request, KV_WINDOW_TILE, format_bytes).
- kv_planner.py (~470 lines): build_kv_cache(), the fitting/grouping
  algorithm (plan_kv_groups + helpers), and KVEventLog.

kv_planner.py re-exports everything via __all__, so the 14 existing call
sites (`from mirage.mpk.kv_planner import ...`) are untouched. Dependency
direction is one-way (kv_planner -> kv_cache/kv_stream), no cycles.

Renamed for clarity, all mechanical (no behavior change):
- KVKind -> KVMode, KVCachePlan -> KVCache, KVGroupSpec -> KVGroupConfig
- KVCache.Group.spec_name -> stream_name (it holds a stream name, not a spec)
- _fit_to_anchor/_fit_block_size -> _unify_to_page_size/_pack_to_page_size,
  matching vLLM's unify_kv_cache_spec_page_size naming
- _warn_if_group_size_starved -> _warn_if_slots_starved

Removed dead code: the __main__ test runner in test_kv_planner.py (defined
mid-file, silently skipped every test declared after it), an unused
Sequence import, and KVCachePlan.merged_with() (zero callers anywhere).

test_kv_planner.py: cut real-model-builder coupling from 12 tests down to
1 (test_gpt_oss_real_config_plan, kept as the sole real-declaration
integration smoke test); the rest now exercise the planner with synthetic
specs so a model builder changing shape can't break planner tests.
Deleted 3 tests that were actually validating a builder's own logic
(deepseek_v3 x2, dflash) rather than the planner. Renamed the
Kimi-Linear-branded bounded-state test to a generic one not tied to an
unimplemented model. Aggressively trimmed docstrings/comments throughout.

61 -> 56 tests, all passing. Every model builder (gpt_oss/deepseek_v3/
dflash/inkling/glm4_moe) still imports and constructs cleanly against the
new module layout.
